requirements.txt
streamlit>=1.28.0
requests>=2.28.0
pypdf>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.0.0
